Remove potentially confusing mentions of commercial database systems
We evaluated our system with an array of commercial systems. E.g., HyPer, Umbra, HANA, Actian Vectorwise and other. We are not going to disclose what DBMS X is due to the DeWitt clause.
This PR removes outdated comments that mentioned other DBMSs that have been used for internal comparison.